requirements.txt
attrs==23.1.0
certifi==2023.7.22
cffi==1.15.1
exceptiongroup==1.1.2
h11==0.14.0
html2text==2020.1.16
idna==3.4
lxml==4.9.3
outcome==1.2.0
pycparser==2.21
PySocks==1.7.1
python-docx==0.8.11
selenium==4.11.2
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
trio==0.22.2
trio-websocket==0.10.3
urllib3==2.0.4
wsproto==1.2.0
